refactor: remove re-export facades after consumer migration
