+++ src/__tests__/components/customizer/armor/variants/CleanTechDiagram.test.tsx
@@ -2,8 +2,8 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import { ARMOR_STATUS } from '@/components/customizer/armor/shared/ArmorFills';
 import { CleanTechDiagram } from '@/components/customizer/armor/variants/CleanTechDiagram';
+import { ARMOR_STATUS } from '@/constants/armorStatus';
 import { MechLocation } from '@/types/construction';
 
 describe('CleanTechDiagram', () => {
