+++ src/__tests__/unit/types/gameplay/TerrainTypes.test.ts
@@ -9,7 +9,7 @@ import {
   ITerrainFeature,
   IHexTerrain,
   TERRAIN_PROPERTIES,
-} from '../../../../types/gameplay/TerrainTypes';
+} from '@/types/gameplay/TerrainTypes';
 
 describe('TerrainTypes', () => {
   describe('TerrainType enum', () => {
