+++ src/components/customizer/armor/variants/MegaMekDiagram.tsx
@@ -12,9 +12,11 @@
 
 import React, { useState } from 'react';
 
+import type { LocationArmorData } from '@/types/construction/LocationArmorData';
+import type { MechConfigType } from '@/types/construction/MechConfigType';
+
 import { MechLocation } from '@/types/construction';
 
-import { LocationArmorData } from '../ArmorDiagram';
 import { ArmorStatusLegend, ArmorDiagramInstructions } from '../shared';
 import {
   getMegaMekStatusColor,
@@ -28,7 +30,6 @@ import { DiagramHeader } from '../shared/DiagramHeader';
 import {
   useResolvedLayout,
   ResolvedPosition,
-  MechConfigType,
   getLayoutIdForConfig,
 } from '../shared/layout';
 import { getLocationLabel, hasTorsoRear } from '../shared/MechSilhouette';
