+++ src/components/customizer/armor/variants/QuadArmorDiagram.tsx
@@ -1,9 +1,10 @@
 import React, { useState } from 'react';
 
+import type { LocationArmorData } from '@/types/construction/LocationArmorData';
+
 import { ArmorDiagramVariant } from '@/stores/useAppSettingsStore';
 import { MechLocation } from '@/types/construction';
 
-import { LocationArmorData } from '../ArmorDiagram';
 import { ArmorDiagramQuickSettings } from '../ArmorDiagramQuickSettings';
 import { GradientDefs } from '../shared/ArmorFills';
 import {
