+++ src/components/customizer/armor/variants/TripodArmorDiagram.tsx
@@ -1,10 +1,11 @@
 import React, { useState } from 'react';
 
+import type { LocationArmorData } from '@/types/construction/LocationArmorData';
+
 import { ArmorDiagramVariant } from '@/stores/useAppSettingsStore';
 import { MechLocation } from '@/types/construction';
 import { TRIPOD_LOCATIONS } from '@/types/construction/MechConfigurationSystem';
 
-import { LocationArmorData } from '../ArmorDiagram';
 import { ArmorDiagramQuickSettings } from '../ArmorDiagramQuickSettings';
 import { GradientDefs } from '../shared/ArmorFills';
 import {
